Backend/requirements.txt
Django == 5.0.3
djangorestframework == 3.14.0
pymongo == 4.7.1
django-cors-headers == 4.3.1
bcrypt == 4.1.3
pycryptodome == 3.20.0
cryptography == 42.0.7